sam/api-pattern2/storefinder-datageneration/requirements.txt
aws-psycopg2
pandas
boto3